chore: Rename test file
